Remove Blueberry component library (now theme only)
