--- src/blueberry/Table.js
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export default props => <table className="bb-table" {...props} />
