+++ src/themes/blueberry/ListItem.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import F from 'futil-js'
 import { observer } from 'mobx-react'
-import { useLens } from '../utils/react'
+import { useLens } from '../../utils/react'
 
 let ListItem = props => {
   let hovering = useLens(false)
